修复一个bug，在设置快捷方式页面选中快捷方式再取消app会报错的问题

+++ app/src/main/java/com/egguncle/xposednavigationbar/ui/activity/SetFunActivity.java
@@ -164,11 +164,11 @@ public class SetFunActivity extends BaseActivity {
                         .setMultiChoiceItems(funcs, null, new DialogInterface.OnMultiChoiceClickListener() {
                             @Override
                             public void onClick(DialogInterface dialog, int which, boolean isChecked) {
-                              //  String name = funcs[which];
+                                //String name = funcs[which];
                                 if (isChecked) {
                                     selectList.add(which);
                                 } else {
-                                    selectList.remove(which);
+                                    selectList.remove((Integer)which);
                                 }
                             }
                         }).create();
